+++ R/plot.R
@@ -24,31 +24,51 @@ ggplot2::autoplot
 #' library(ggridges)
 #'
 #' # Load in posterior parameter example
-#' data(posterior_param_ex)
+#' data(comp_bayes_lm_ex)
 #'
 #' # Plot beta_0, growth intercepts
-#' autoplot(posterior_param_ex, type = "intercepts")
+#' autoplot(comp_bayes_lm_ex, type = "intercepts")
 #'
 #' # Plot beta_dbh, growth-dbh slope
-#' autoplot(posterior_param_ex, type = "dbh_slopes")
+#' autoplot(comp_bayes_lm_ex, type = "dbh_slopes")
 #'
 #' # Plot lambdas, competition coefficients
-#' autoplot(posterior_param_ex, type = "competition")
+#' autoplot(comp_bayes_lm_ex, type = "competition")
 #' @export
 autoplot.comp_bayes_lm <- function(object,
                                  type = "intercepts",
                                  sp_to_plot = NULL,
                                  ...) {
 
+  check_comp_bayes_lm(object)
+  check_inherits(type, "character")
+  if (!type %in% c("intercepts", "dbh_slopes", "competition")) {
+    glue_stop('The `type` argument must be one of "intercepts", ',
+              '"dbh_slopes", or "competition".')
+  }
+
+
   # Identify all species and baseline category of species used for regression
-  sp_list <- object$sp_list
+  sp_list <- object$post_params$sp_list
   baseline_species <- sp_list %>%
     sort() %>%
     .[1]
 
+  if (!is.null(sp_to_plot)) {
+    check_inherits(sp_to_plot, "character")
+
+    if (any(!sp_to_plot %in% sp_list)) {
+      glue_stop("The `sp_to_plot` argument must be a subset of species in the ",
+                "training data. \nThe following elements of `sp_to_plot` could ",
+                "not be matched to a ",
+                "species: {list(sp_to_plot[!sp_to_plot %in% sp_list])}\n"
+      )
+    }
+  }
+
   # Simulate observations from posterior
   beta_lambda_posterior_df <- simulate_beta_lambda_posterior(
-    object,
+    object$post_params,
     sp_list,
     baseline_species
   )