+++ data-raw/process_data_raw.R
@@ -153,9 +153,9 @@ use_data(focal_vs_comp_ex, overwrite = TRUE)
 
 
 # fit the model
-posterior_param_ex <- focal_vs_comp_ex %>%
+comp_bayes_lm_ex <- focal_vs_comp_ex %>%
   comp_bayes_lm()
-use_data(posterior_param_ex, overwrite = TRUE)
+use_data(comp_bayes_lm_ex, overwrite = TRUE)